divided the code in index.js in sub files so it is easier to read
