+++ app/options.js
@@ -1,0 +1,9 @@
+// define below the buzz options
+export const buzzOptions = {
+    0: [],
+    1: [9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21],
+    2: [9, 11, 13, 15, 17, 19, 21],
+    3: [9, 12, 15, 18, 21]
+};
+
+export const production = true
